[WIP] Refactored actions and reducers. Figuring out Redux slowly but surely
