--- src/actions/actionTypes.ts
@@ -1,2 +1,0 @@
-// TODO Define action types
-export const TOGGLE_CATEGORY = "TOGGLE_CATEGORY"
