--- src/actions/index.ts
@@ -1,8 +1,0 @@
-import {TOGGLE_CATEGORY} from "./actionTypes"
-
-// TODO Define actions
-export const toggleCategoryAction = () => {
-  return {
-    type: TOGGLE_CATEGORY
-  }
-}
