+++ src/store/rootReducer.ts
@@ -1,8 +1,8 @@
-import blocksCategoryReducer from "./blocksCategoryReducer"
+import blocksCategoryReducer from "./features/blocks/blocksCategory"
 import {combineReducers} from "redux"
 
 const rootReducer = combineReducers({
-  blocksCategoryReducer: blocksCategoryReducer
+  blocksCategoryReducer
 })
 
 export default rootReducer
